requirements.txt
# Core dependencies
python-telegram-bot>=20.0
asyncio>=3.4.3
psutil>=5.9.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Performance monitoring
prometheus-client>=0.14.0
statsd>=3.3.0

# Utility packages
python-dotenv>=0.19.0
pydantic>=1.9.0 